myfirst/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-unfold==0.22.0
sqlparse==0.5.0
typing_extensions==4.11.0
